requirements.txt
aws-cdk-lib==2.150.0
fastapi==0.111.1
requests==2.32.3
